+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Changed
+- Add `min-width` and `text-align` for Dropdown's button.
+- Add `isFluid` prop to `Dropdown` and `ToggleMenu` components.
+
 ## [0.8.2] - 2018-09-27
 ### Changed
 - Fix classnames in Modal component